+++ app/src/main/java/bcc/app/java/AppUsingBlockchain.java
@@ -18,18 +18,25 @@ public class AppUsingBlockchain {
 		List<String> nodes = new LinkedList<>();
 		// geth --datadir dev --dev --rpc --rpcapi "eth,personal,net,txpool" --rpcport 61337 --ipcdisable console
 		// 65164
-		nodes.add("http://localhost:61337");
-		String account ="0x315C4BBbD11bA58ad0c52D25F3035ca810d9e6D1";
+//		nodes.add("http://localhost:61337");
+//		String account ="0x315C4BBbD11bA58ad0c52D25F3035ca810d9e6D1";
+		
+		nodes.add("http://localhost:8501");
+		String account ="0x128ad75a1755e23c989ae00d099a18b628bc1f11";
+		
 		List<String> accounts = new LinkedList<>();
 		accounts.add(account);
 		
 		EthereumBlockchain ebc = new EthereumBlockchain(nodes, accounts);
 		Blockchain bc = ebc;
+		//bc.setTraceHttpRequestResponce(true);
 		
 		int trCount = ebc.eth_getTransactionCount(account);
 		log.info("{}", trCount);
 		
 		BCTransaction tr = new BCTransaction();
+		tr.setFrom(account);
+		tr.setTo(account);
 		String trid = bc.saveTransaction(tr);
 
 		BCTransaction tr2 = bc.readTransaction(trid);