+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 
+### 0.2 
+- option to trace HTTP request/responce
+Look at
+- Bitcoin
+- Ethereum Classic
+
 ### 0.1 2018-4-25
 - can save transactions